WIP: Start on a fake implementation of live-kit
