+++ crates/live_kit_client/Cargo.toml
@@ -11,6 +11,9 @@ doctest = false
 [[example]]
 name = "test_app"
 
+[features]
+test-support = []
+
 [dependencies]
 media = { path = "../media" }
